--- src/main/java/me/daeho/step3/README.md
@@ -1,11 +1,0 @@
-## 4단계 - 자동차 경주
-### 구현 기능
- - Car
-   1. 준비 : 차 이름, 현재 위치 0으로 초기화 및 전진 룰 셋팅 
-   2. 이동 : 전진 룰에 부합하면 전진
-   3. 확인 : 현재 위치 확인, 이름 확인
- - RacingGame
-   1. 초기화 : 참여 자동차 세팅, 경주의 전진 룰 주입
-   2. 횟수 확인 : 이동 가능 횟수가 남아있는지 확인
-   3. 횟수 차감 : 이동 후 이동한 횟수를 차감한다.
-   4. 1등 확인 : 이동 시 1등의 위치 기록, 경주 종료 후 우승 자동차 확인
