+++ src/main/java/me/daeho/step5/README.md
@@ -1,0 +1,5 @@
+## 5단계 - 자동차 경주
+### 구현 기능
+ - Package 구조 변경 
+   1. RacingCar, Car 객체 domain 패키지로 이동
+   2. InputView, ResultView 객체 view 패키지로 이동
