+++ src/main/java/me/daeho/step5/view/InputView.java
@@ -1,4 +1,4 @@
-package me.daeho.step3.ui;
+package me.daeho.step5.view;
 
 import me.daeho.utils.StringUtils;
